requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
python-dotenv
pydantic
pydantic-settings
requests
